css: globally decrease font size
We frequently run out of space on Sciware slides. Even accounting for the fact
that readability is extremely important, we could almost certainly use a smaller
font size and still have a very reasonable content density.

We might tweak this value once we see it on the projector screens, but let's start
with this.